21w17a

+++ net/minecraft/item/map/MapIcon.java
@@ -132,7 +132,7 @@ public class MapIcon {
         }
 
         public static Type byId(byte id) {
-            return Type.values()[MathHelper.clamp(id, 0, Type.values().length - 1)];
+            return Type.values()[MathHelper.clamp((int)id, 0, Type.values().length - 1)];
         }
     }
 }
