+++ net/minecraft/network/packet/c2s/play/UpdateStructureBlockC2SPacket.java
@@ -57,9 +57,9 @@ implements Packet<ServerPlayPacketListener> {
         this.mode = buf.readEnumConstant(StructureBlockMode.class);
         this.structureName = buf.readString();
         int i = 48;
-        this.offset = new BlockPos(MathHelper.clamp(buf.readByte(), -48, 48), MathHelper.clamp(buf.readByte(), -48, 48), MathHelper.clamp(buf.readByte(), -48, 48));
+        this.offset = new BlockPos(MathHelper.clamp((int)buf.readByte(), -48, 48), MathHelper.clamp((int)buf.readByte(), -48, 48), MathHelper.clamp((int)buf.readByte(), -48, 48));
         int j = 48;
-        this.size = new Vec3i(MathHelper.clamp(buf.readByte(), 0, 48), MathHelper.clamp(buf.readByte(), 0, 48), MathHelper.clamp(buf.readByte(), 0, 48));
+        this.size = new Vec3i(MathHelper.clamp((int)buf.readByte(), 0, 48), MathHelper.clamp((int)buf.readByte(), 0, 48), MathHelper.clamp((int)buf.readByte(), 0, 48));
         this.mirror = buf.readEnumConstant(BlockMirror.class);
         this.rotation = buf.readEnumConstant(BlockRotation.class);
         this.metadata = buf.readString(12);
